Move the resampling actor to the `microgrid` module
